#11: Refactoring
