+++ constants/Color.tsx
@@ -16,4 +16,11 @@ const Color = {
   },
 };
 
-export default Color;
+const generateColor = ():string => {
+  const obj = Color.primary.slots;
+  const keys = Object.keys(obj);
+  const color = keys[Math.floor(Math.random() * keys.length)];
+  return obj[color];
+};
+
+export {Color,generateColor};
