+++ views/Game/index.tsx
@@ -1,19 +1,11 @@
 import { View, Text,SafeAreaView, StyleSheet,TouchableOpacity } from "react-native";
-import { Color, Screen, Player } from "../../constants";
+import { generateColor, Screen, Player } from "../../constants";
 import { useEffect, useState, useRef, forwardRef } from "react";
-import { Audio } from 'expo-av';
 import { MathHelper } from '../../helpers';
-import { usePrevious } from "../../hooks";
 import * as Animatable from 'react-native-animatable';
 import {failSound,stopSound,startSound,normalSound,clockSound,pressSound} from "../../helpers/SoundPlayer";
+import {PrecentageBar, ColorPad} from "../../components";
 
-
-const generateColor = ():string => {
-  const obj = Color.primary.slots;
-  const keys = Object.keys(obj);
-  const color = keys[Math.floor(Math.random() * keys.length)];
-  return obj[color];
-};
 
 const CountDown = ({active,countDownRef, countDown}) => {
   return (
@@ -22,37 +14,6 @@ const CountDown = ({active,countDownRef, countDown}) => {
       <Animatable.Text duration={2000} iterationCount={4} style={{fontSize: 300,color: "white"}}>{countDown}</Animatable.Text>
     </Animatable.View>
   </View>
-  )
-}
-
-const PrecentageBar = ({value,color,onPressStop, player, gameState}:any) => {
-  const animatableRef = useRef<Animatable.View & View>(null);
-  const prevColor = usePrevious(color);
-  const prevValue = usePrevious(value);
-  useEffect(() => {
-    if(prevValue !== 100){
-      animatableRef.current && animatableRef.current.animate({0: {backgroundColor: prevColor, width: (100 - (prevValue || 0)).toString()+"%"},1: {backgroundColor: color, width: (100 - value).toString()+"%"}});
-    }
-  },[color,value]);
-
-
-  const styles = StyleSheet.create({
-    wrapper: {position: "relative",width: Screen.Width,height: 30,backgroundColor:color},
-  });
-
-  return (
-    <View style={{width: Screen.Width, display: "flex",alignItems: "center",justifyContent: player == Player.One ? "space-between" : "center"}}>
-    { value !== 100 ? (
-      
-    <Animatable.View duration={250} ref={animatableRef} style={styles.wrapper}/>
-    ) : (
-    <TouchableOpacity onPress={onPressStop} style={{opacity: gameState.blocked === player ? 0.5 : 1}}>
-    <Animatable.View duration={500} animation="flipInX" style={{display: "flex",justifyContent:"center",alignItems:"center",width: Screen.Width * 0.3,height: Screen.Height * 0.07,backgroundColor: color, borderRadius: 10}}>
-      <Animatable.Text duration={1000} animation="pulse" iterationCount="infinite" style={{color:"white",fontSize: 30}}>STOP</Animatable.Text>
-    </Animatable.View>
-    </TouchableOpacity>
-    )}
-    </View>
   )
 }
 
@@ -142,85 +103,6 @@ const PlayerController = (props) => {
             </Animatable.View>
           </View>
         </View>
-  );
-};
-
-const ColorPad = (props) => {
-  const { _onClick, active, player, gameState } = props;
-
-  const [color, setColor] = useState(gameState.mode === "oneColor" ? "black" : generateColor());
-  const [time, setTime] = useState(0);
-  const [seed,setSeed] = useState(0);
-  const animatableRef = useRef<Animatable.View & View>(null); 
-  
-  const getSeedColor = () => {
-    if(seed == 0){
-      setSeed(1);
-      return Color.primary.slots.slot_01;
-    } else if(seed == 1){
-      setSeed(2);
-      return Color.primary.slots.slot_02;
-    } else if(seed == 2){
-      setSeed(3);
-      return Color.primary.slots.slot_03;
-    } else if(seed == 3){
-      setSeed(0);
-      return Color.primary.slots.slot_04;
-    }
-  }
-  const updateColor = () => {
-    const newColor = gameState.mode == "oneColor" ? getSeedColor() : generateColor();
-
-      if(newColor === color && gameState.mode !== "oneColor") {
-        updateColor();
-      }
-      setColor(newColor);
-    animatableRef.current.animate({0: {backgroundColor: color},1: {backgroundColor: newColor}})
-
-  }
-  useEffect(() => {
-    let intervall = setInterval(() => {
-      if (time == 0) {
-        updateColor();
-        setTime(gameState.mode == "oneColor" ? 1 : MathHelper.randomIntFromInterval(1, 3));
-      } else {
-        setTime((prevTime) => prevTime - 1);
-      }
-    }, 1000);
-    return () => {
-      clearInterval(intervall);
-    };
-  }, [time]);
-
-  const onPress = ():void => {
-    if(!active){
-      return;
-    }
-    animatableRef.current.pulse();
-    setColor(generateColor());
-    setTime(gameState.mode == "oneColor" ? 1 : MathHelper.randomIntFromInterval(1, 5));
-    _onClick(color,player);
-  };
-  return (
-    <TouchableOpacity onPress={onPress}>
-      <Animatable.View ref={animatableRef} duration={250}
-        style={{
-          width: Screen.Width * 0.2,
-          height: Screen.Width * 0.2,
-          backgroundColor: color,
-          borderRadius: 10,
-          display: "flex",
-          alignItems: "center",
-          justifyContent: "center",
-          opacity: gameState.blocked === player ? 0.5 : 1
-        }}
-      >
-        {gameState.blocked === player && (
-            <Text style={{fontSize: 50,color: "white"}}>⨉</Text>
-        )}
-    </Animatable.View>
-
-    </TouchableOpacity>
   );
 };
